feat: Add deleteOldImage function to cloudinary utils and used it in userController and some minor changes to subscription and user models
This commit adds the `deleteOldImage` function to the `cloudinary.js` file in the `utils` directory. The function handles the logic for deleting an image from Cloudinary using the `destroy` method. It takes the `public_id` of the image as a parameter and deletes the corresponding image from the Cloudinary storage. This addition enhances the functionality of the application by providing a way to delete images from Cloudinary.